Fix expanding mem to also do rehashing
Whenever capacity changes, we need to re-calculate the hash function.
This will reduce the chance of collision gradually as the capacity
increases.

Prior to rehashing operation, we will need double of memory space of
hashmap according to its new size before it reduces down to just half of
it at the end of the process as we need to allocate new memory space
before re-calculate hashing to add new elements again.