Add custom memory freeing handler for element in hashmap, allow user to add function pointer
Added custom function pointer allowing user to add their own function to
free memory space for elements when time of deletion, clearing, and
rehashing come.

Notice when we do rehashing, we don't call user's custom freeing
function for elements as we just copy the memory space of element into a
new container; new hashmap's `mem` memory space. If we do free it as
well, then pointer is orphan later; it's freed but we still keep address
of it. With this, see `free_mem_except_elems()` local function.